UX-513 Resolve circular dependency
